refactor(withLocales): Colocate hoc together
