+++ packages/cozy-sharing/src/ConfirmTrustedRecipientsDialog.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { default as DumbConfirmTrustedRecipientsDialog } from './components/ConfirmTrustedRecipientsDialog'
 import SharingContext from './context'
-import withLocales from './withLocales'
+import withLocales from './hoc/withLocales'
 
 export const ConfirmTrustedRecipientsDialog = withLocales(
   ({ document, ...rest }) => (
