+++ packages/cozy-sharing/src/SharedBadge.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { default as DumbSharedBadge } from './components/SharedBadge'
 import SharingContext from './context'
-import withLocales from './withLocales'
+import withLocales from './hoc/withLocales'
 
 export const SharedBadge = withLocales(({ docId, ...rest }) => (
   <SharingContext.Consumer>
